Refactor search and list commands to playlist module
Moved `search` and `list` business logic from `src/main.py` to `src/commands/playlist.py` to adhere to Single Responsibility Principle.
Updated `tests/unit/test_main.py` to mock consoles in the new location.
