+++ src/main.py
@@ -1,11 +1,15 @@
 import typer
 import sys
-import concurrent.futures
 from pathlib import Path
 from typing import Optional
 from rich.console import Console
 from .spotify_client import get_spotify
-from .commands.playlist import move_tracks as do_move_tracks
+from .commands.playlist import (
+    move_tracks as do_move_tracks,
+    add_tracks as do_add_tracks,
+    search_tracks,
+    list_playlist_tracks
+)
 
 
 app = typer.Typer(help="Swedish Army Knife for Spotify actions.")
@@ -58,29 +62,6 @@ def move(
     except Exception as e:
         err_console.print(f"[bold red]Move Failed:[/] {str(e)}")
 
-def _search_worker(sp, line: str):
-    line = line.strip()
-    if not line:
-        return None
-
-    # Parse "Artist - Title" format
-    if " - " not in line:
-        err_console.print(f"[yellow]Skipping invalid format:[/] {line}")
-        return None
-
-    artist, title = line.split(" - ", 1)
-    try:
-        result = sp.search(q=f'artist:{artist} track:{title}', type='track', limit=1)
-
-        if result['tracks']['items']:
-            return result['tracks']['items'][0]
-        else:
-            err_console.print(f"[red]Not found:[/] {artist} - {title}")
-            return None
-    except Exception as e:
-        err_console.print(f"[red]Error searching for:[/] {line} - {str(e)}")
-        return None
-
 @playlist_app.command(name="search")
 def search(
     output: str = typer.Option("uri", "--output", "-o", help="Output format: uri (default), id, text")
@@ -98,22 +79,8 @@ def search(
     
     lines = sys.stdin.readlines()
 
-    # Use ThreadPoolExecutor for parallel processing
-    # Limit workers to avoid rate limits
-    with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
-        # Submit all tasks and preserve order
-        futures = [executor.submit(_search_worker, sp, line) for line in lines]
-        
-        for future in futures:
-            track = future.result()
-            if track:
-                if output == "id":
-                    print(track['id'])
-                elif output == "text":
-                    artists = ', '.join([a['name'] for a in track['artists']])
-                    print(f"{artists} - {track['name']}")
-                else:
-                    print(track['uri'])
+    # Delegate logic to command module
+    search_tracks(sp, lines, output)
 
 @playlist_app.command(name="list")
 def list_tracks(
@@ -137,19 +104,7 @@ def list_tracks(
     
     try:
         sp = get_spotify()
-        results = sp.playlist_tracks(playlist_id)
-        while results:
-            for item in results['items']:
-                track = item['track']
-                if track:  # Can be None for local/unavailable tracks
-                    if output == "id":
-                        print(track['id'])
-                    elif output == "uri":
-                        print(track['uri'])
-                    else:
-                        artists = ', '.join([a['name'] for a in track['artists']])
-                        print(f"{artists} - {track['name']}")
-            results = sp.next(results) if results.get('next') else None
+        list_playlist_tracks(sp, playlist_id, output)
     except Exception as e:
         err_console.print(f"[bold red]Error:[/] {str(e)}")
         raise typer.Exit(1)
@@ -162,7 +117,6 @@ def add_tracks_to_playlist(
 ):
     """Add tracks to a playlist. Reads track URIs from file or stdin."""
     import re
-    from .commands.playlist import add_tracks as do_add_tracks
 
     # 1. Resolve Playlist ID
     if url:
@@ -204,4 +158,3 @@ def add_tracks_to_playlist(
 
 if __name__ == "__main__":
     app()
-
